refactor(docx): extract float-wrap geometry to float-layout.ts
Move the pure, dependency-free float-wrap math out of the 4.8k-line
renderer into a sibling module: FloatRect / Gap types, the tolerance
constants (FLOAT_OVERLAP_EPS / FLOAT_PAGE_RIGHT_SLACK / MIN_LINE_GAP),
isWrapFloat, rectsOverlap, widestFreeGap, resolveLineFloatWindow,
resolveFloatOverlap and skipPastTopAndBottom.

resolveFloatOverlap previously took the whole RenderState but only read
pageWidth/scale/floats; its signature now takes pageRight and the float
array directly, so the new module has zero dependency on renderer
internals (no circular import). The drawing-heavy helpers
(registerAnchorFloats / renderAnchorImages / renderEmptyMarkParagraph /
resolveAnchorBox) stay in renderer.ts and import the geometry.

Pure code motion. typecheck + build pass. VRT 21/21, demo/sample-1 100%,
sample-9 all 11 pages pixel-identical.
